requirements.txt
Package                       Version
----------------------------- -----------
absl-py                       2.1.0
aiosmtplib                    2.0.2
alembic                       1.13.1
annotated-types               0.7.0
anyio                         4.4.0
astunparse                    1.6.3
azure-ai-vision-imageanalysis 1.0.0b2
azure-core                    1.30.1
bcrypt                        4.1.2
blinker                       1.7.0
certifi                       2024.8.30
cffi                          1.16.0
charset-normalizer            3.3.2
click                         8.1.7
colorama                      0.4.6
contourpy                     1.3.0
cryptography                  42.0.5
cycler                        0.12.1
dnspython                     2.6.1
easyocr                       1.7.1
ecdsa                         0.18.0
editdistance                  0.8.1
efficientnet                  1.0.0
email-validator               2.1.0.post1
essential-generators          1.0
fastapi                       0.114.1
fastapi-mail                  1.4.1
filelock                      3.16.0
flatbuffers                   24.3.25
fonttools                     4.53.1
fsspec                        2024.9.0
gast                          0.6.0
google-pasta                  0.2.0
greenlet                      3.0.3
grpcio                        1.66.1
h11                           0.14.0
h5py                          3.11.0
httpcore                      1.0.3
httptools                     0.6.1
httpx                         0.26.0
idna                          3.8
imageio                       2.35.1
imgaug                        0.4.0
isodate                       0.6.1
itsdangerous                  2.1.2
Jinja2                        3.1.4
keras                         3.5.0
Keras-Applications            1.0.8
keras-ocr                     0.9.3
kiwisolver                    1.4.7
lazy_loader                   0.4
libclang                      18.1.1
lmdb                          1.5.1
logtail-python                0.2.10
Mako                          1.3.2
Markdown                      3.7
markdown-it-py                3.0.0
MarkupSafe                    2.1.5
matplotlib                    3.9.2
mdurl                         0.1.2
ml-dtypes                     0.4.0
mpmath                        1.3.0
msgpack                       1.0.8
namex                         0.0.8
networkx                      3.3
ninja                         1.11.1.1
numpy                         1.26.4
opencv-python                 4.10.0.84
opencv-python-headless        4.10.0.84
opt-einsum                    3.3.0
optree                        0.12.1
orjson                        3.9.14
packaging                     24.1
pandas                        2.2.2
passlib                       1.7.4
pillow                        10.4.0
pip                           24.2
protobuf                      4.25.4
psutil                        6.0.0
psycopg2                      2.9.9
psycopg2-binary               2.9.9
py-cpuinfo                    9.0.0
pyasn1                        0.5.1
pyclipper                     1.3.0.post5
pycparser                     2.21
pydantic                      2.9.1
pydantic_core                 2.23.3
pydantic-extra-types          2.5.0
pydantic-settings             2.2.1
Pygments                      2.18.0
pyparsing                     3.1.4
pytesseract                   0.3.13
python-bidi                   0.6.0
python-dateutil               2.9.0.post0
python-dotenv                 1.0.1
python-jose                   3.3.0
python-multipart              0.0.9
pytz                          2024.2
PyYAML                        6.0.2
rapidfuzz                     3.9.7
requests                      2.32.3
rich                          13.8.1
rsa                           4.9
scikit-image                  0.24.0
scipy                         1.14.1
seaborn                       0.13.2
setuptools                    65.5.0
shapely                       2.0.6
six                           1.16.0
sniffio                       1.3.1
soupsieve                     2.6
SQLAlchemy                    2.0.27
starlette                     0.38.5
sympy                         1.13.2
tensorboard                   2.17.1
tensorboard-data-server       0.7.2
tensorflow                    2.17.0
tensorflow-intel              2.17.0
tensorflow-io-gcs-filesystem  0.31.0
termcolor                     2.4.0
tifffile                      2024.8.30
torch                         2.4.1
torchvision                   0.19.1
tqdm                          4.66.5
typing_extensions             4.12.2
tzdata                        2024.1
ujson                         5.9.0
ultralytics                   8.2.91
ultralytics-thop              2.0.6
urllib3                       2.2.2
uuid                          1.30
uvicorn                       0.30.6
validators                    0.34.0
watchfiles                    0.21.0
websockets                    12.0
Werkzeug                      3.0.4
wheel                         0.44.0
wrapt                         1.16.0
